+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.52.1] - 2026-02-05
 
 ### Added